requirements.txt
requests
pytest
jsonschema
selene==2.0.0rc9
allure-pytest